Revert "Move to v141 compile tools"
This reverts commit 7b27e23a3f561f0f037634f8e7af1863fb2f3e1c.

+++ src/game/client/c_impact_effects.h
@@ -36,12 +36,12 @@ public:
 
 #ifdef _XBOX
 		//Cap the minimum roll
-		if ( fabsf( pParticle->m_flRollDelta ) < 0.1f )
+		if ( fabs( pParticle->m_flRollDelta ) < 0.1f )
 		{
 			pParticle->m_flRollDelta = ( pParticle->m_flRollDelta > 0.0f ) ? 0.1f : -0.1f;
 		}
 #else
-		if ( fabsf( pParticle->m_flRollDelta ) < 0.5f )
+		if ( fabs( pParticle->m_flRollDelta ) < 0.5f )
 		{
 			pParticle->m_flRollDelta = ( pParticle->m_flRollDelta > 0.0f ) ? 0.5f : -0.5f;
 		}
