+++ src/game/client/particle_simple3d.cpp
@@ -63,21 +63,21 @@ void CSimple3DEmitter::SimulateParticles( CParticleSimulateIterator *pIterator )
 				pParticle->m_vAngles.x = anglemod(pParticle->m_vAngles.x);
 				if (pParticle->m_vAngles.x < 180)
 				{
-					if (fabsf(pParticle->m_vAngles.x - 90) > 0.5f)
+					if (fabs(pParticle->m_vAngles.x - 90) > 0.5)
 					{
 						pParticle->m_vAngles.x = 0.5*pParticle->m_vAngles.x + 46;
 					}
 				}
 				else
 				{
-					if (fabsf(pParticle->m_vAngles.x - 270) > 0.5f)
+					if (fabs(pParticle->m_vAngles.x - 270) > 0.5)
 					{
 						pParticle->m_vAngles.x = 0.5*pParticle->m_vAngles.x + 135;
 					}
 				}
 
 				pParticle->m_vAngles.y = anglemod(pParticle->m_vAngles.y);
-				if (fabsf(pParticle->m_vAngles.y) > 0.5f)
+				if (fabs(pParticle->m_vAngles.y) > 0.5)
 				{
 					pParticle->m_vAngles.y = 0.5*pParticle->m_vAngles.z;
 				}
